# packages/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/iam/supervisors": {
      "get": {
        "operationId": "IamController_listSupervisors",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SupervisorResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Iam"
        ]
      },
      "post": {
        "operationId": "OperarioController_createSupervisor",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSupervisorBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Returns the created Supervisor.id (not the User.id).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedIdDto"
                }
              }
            }
          }
        },
        "tags": [
          "Operario"
        ]
      }
    },
    "/iam/supervisors/{id}": {
      "get": {
        "operationId": "IamController_getSupervisor",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupervisorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Iam"
        ]
      }
    },
    "/iam/operarios": {
      "get": {
        "operationId": "IamController_listOperarios",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OperarioResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Iam"
        ]
      },
      "post": {
        "operationId": "OperarioController_createOperario",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOperarioBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedIdDto"
                }
              }
            }
          }
        },
        "tags": [
          "Operario"
        ]
      }
    },
    "/iam/operarios/{id}": {
      "get": {
        "operationId": "IamController_getOperario",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperarioResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Iam"
        ]
      },
      "patch": {
        "operationId": "OperarioController_reassignOperario",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReassignOperarioBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperarioResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Operario"
        ]
      }
    },
...
    "/iam/assignments/{id}": {
      "get": {
        "operationId": "IamController_getAssignment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Iam"
        ]
      }
    },
...
  },
  "info": {
    "title": "FuturaGest API",
    "description": "",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SupervisorResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "municipioId": {
            "type": "string",
            "format": "uuid"
          },
          "zoneId": {
            "type": "string",
            "format": "uuid"
          },
          "area": {
            "type": "string",
            "description": "SupervisorArea: BARRIDO | RECOLECCION | SUPERNUMERARIO"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "createdAt": {
            "type": "string",
            "description": "ISO 8601 timestamp"
          }
        },
        "required": [
          "id",
          "userId",
          "municipioId",
          "zoneId",
          "area",
          "email",
          "createdAt"
        ]
      },
      "OperarioResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fullName": {
            "type": "string"
          },
          "documento": {
            "type": "string"
          },
          "supervisorId": {
            "type": "string",
            "format": "uuid"
          },
          "cargo": {
            "type": "string",
            "description": "Free-text job position (e.g. \"Barrido\", \"Recolección\")"
          },
          "active": {
            "type": "boolean",
            "description": "Derived: deactivatedAt === null"
          },
          "deactivatedAt": {
            "type": "object",
            "description": "ISO 8601 timestamp or null when active",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "description": "ISO 8601 timestamp"
          },
          "updatedAt": {
            "type": "string",
            "description": "ISO 8601 timestamp — delta cursor for ?since= queries"
          }
        },
        "required": [
          "id",
          "fullName",
          "documento",
          "supervisorId",
          "cargo",
          "active",
          "deactivatedAt",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreatedIdDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "id"
        ]
      },
...
      "CreateOperarioBody": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string"
          },
          "documento": {
            "type": "string",
            "description": "Unique national document number"
          },
          "supervisorId": {
            "type": "string",
            "format": "uuid"
          },
          "cargo": {
            "type": "string",
            "description": "Free-text job position (e.g. \"Barrido\", \"Recolección\")",
            "default": ""
          }
        },
        "required": [
          "fullName",
          "documento",
          "supervisorId",
          "cargo"
        ]
      },
...
      "ReassignOperarioBody": {
        "type": "object",
        "properties": {
          "supervisorId": {
            "type": "string",
            "format": "uuid",
            "description": "New supervisor id"
          }
        },
        "required": [
          "supervisorId"
        ]
      },
      "CreateSupervisorBody": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email for the new supervisor user account"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "description": "Temporary password (mustChangePassword=true)"
          },
          "area": {
            "type": "string",
            "enum": [
              "BARRIDO",
              "RECOLECCION",
              "SUPERNUMERARIO"
            ],
            "description": "Operational area: BARRIDO | RECOLECCION | SUPERNUMERARIO"
          },
          "zoneId": {
            "type": "string",
            "format": "uuid",
            "description": "Zone the supervisor is assigned to"
          },
          "municipioId": {
            "type": "string",
            "format": "uuid",
            "description": "Municipio within the zone"
          }
        },
        "required": [
          "email",
          "password",
          "area",
          "zoneId",
          "municipioId"
        ]
      },
...
    }
  }
}